NO-REF: Separting dev setup process and seed data process
